Isosti muutoksia. Poistettu turhia rajapintoja, lisätty yksi uusi. Muotoiltu uudestaan knn-recommenderin alustus
