+++ package.json
@@ -15,7 +15,7 @@
 	},
 	"keywords": [],
 	"author": "",
-	"license": "ISC",
+	"license": "MIT",
 	"bugs": {
 		"url": "https://github.com/bgf122/dataTinderBack/issues"
 	},
@@ -24,12 +24,9 @@
 		"cors": "^2.8.5",
 		"dotenv": "^8.2.0",
 		"express": "^4.17.1",
-		"firebase-admin": "^9.5.0",
 		"knn-recommender": "^3.0.0",
 		"mongoose": "^5.11.17",
-		"morgan": "^1.10.0",
 		"morgan-body": "^2.6.5",
-		"node-fetch": "^2.6.1",
 		"nodemon": "^2.0.7"
 	},
 	"devDependencies": {
