+++ routes/popular.js
@@ -1,8 +1,8 @@
 const express = require('express');
 
 const router = express.Router();
-const service = require('../service/series');
+const service = require('../service/programs');
 
-router.get('/', service.getAllSeries);
+router.get('/', service.getPopularPrograms);
 
 module.exports = router;
