+++ service/programs.js
@@ -1,4 +1,7 @@
 const Program = require('../models/program');
+const User = require('../models/user');
+
+const getSingleProgram = async (program) => Program.findById(program);
 
 exports.getAllPrograms = async (req, res) => {
   try {
@@ -17,3 +20,22 @@ exports.getProgram = async (req, res) => {
     res.json({ error: err.message });
   }
 };
+
+exports.getPopularPrograms = async (req, res) => {
+  console.log("get popular")
+  try {
+    const userData = await User.aggregate([
+      { $unwind: '$data' },
+      { $match: { 'data.value': 1 } },
+      { $group: { _id: '$data.programId', count: { $sum: 1 } } },
+      { $sort: { count: -1 } },
+      { $limit: 10 },
+    ]);
+    const programData = async () => Promise.all(userData.map((program) => getSingleProgram(program._id)));
+
+    return programData().then((data) => res.json(data));
+  } catch (err) {
+    return res.json({ error: err.message });
+  }
+};
+
